requirements.txt
alembic==1.4.2
blinker==1.4
click==7.1.1
dnspython==1.16.0
Flask==1.1.2
Flask-DebugToolbar==0.11.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.19
protobuf==3.11.3
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
redis==3.4.1
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
WTForms==2.2.1
